+++ examples/react-native/storage.js
@@ -3,5 +3,6 @@
  * so that it can be swapped out in the CI to test
  * different storages.
  */
+import { wrappedValidateAjvStorage } from 'rxdb/plugins/validate-ajv';
 import { getRxStorageMemory } from 'rxdb/plugins/storage-memory';
-export const STORAGE = getRxStorageMemory();
+export const STORAGE = wrappedValidateAjvStorage({ storage: getRxStorageMemory() });