+++ src/plugins/storage-lokijs/index.ts
@@ -1,3 +1,0 @@
-export * from './rx-storage-lokijs.ts';
-export * from './lokijs-helper.ts';
-export * from './rx-storage-instance-loki.ts';