+++ test/unit/replication-protocol.test.ts
@@ -299,11 +299,6 @@ useParallel(testContext + ' (implementation: ' + config.storage.name + ')', () =
              * @link https://github.com/pubkey/rxdb/pull/3627
              */
             it('#3627 should not write a duplicate checkpoint', async () => {
-                if (config.storage.name === 'lokijs') {
-                    // TODO fix this with the lokijs storage
-                    return;
-                }
-
                 const masterInstance = await createRxStorageInstance(1);
                 const forkInstance = await createRxStorageInstance(0);
                 const metaInstance = await createMetaInstance(forkInstance.schema);