+++ test/unit/rx-database.test.ts
@@ -22,6 +22,9 @@ import {
 import {
     getRxStorageMemory
 } from '../../plugins/storage-memory/index.mjs';
+import {
+    wrappedValidateAjvStorage
+} from '../../plugins/validate-ajv/index.mjs';
 
 describeParallel('rx-database.test.ts', () => {
     describe('createRxDatabase()', () => {
@@ -100,7 +103,7 @@ describeParallel('rx-database.test.ts', () => {
                 const name = randomCouchString(10);
                 const db = await createRxDatabase({
                     name,
-                    storage: getRxStorageMemory()
+                    storage: wrappedValidateAjvStorage({ storage: getRxStorageMemory() })
                 });
                 const db2 = await createRxDatabase({
                     name,
@@ -110,14 +113,7 @@ describeParallel('rx-database.test.ts', () => {
                 db2.close();
             });
             it('2 password-instances on same adapter', async () => {
-                if (
-                    config.storage.name === 'lokijs' ||
-                    !config.storage.hasMultiInstance
-                ) {
-                    /**
-                     * TODO on lokijs this test somehow fails
-                     * to properly clean up the open broadcast channels.
-                     */
+                if (!config.storage.hasMultiInstance) {
                     return;
                 }
                 const name = randomCouchString(10);
@@ -465,10 +461,6 @@ describeParallel('rx-database.test.ts', () => {
             });
             it('create 2 times on same adapter with different schema', async () => {
                 if (!config.storage.hasMultiInstance) {
-                    return;
-                }
-                if (config.storage.name === 'lokijs') {
-                    // TODO
                     return;
                 }
                 const name = randomCouchString(10);